Replace in-place array changes to support JAX and Tensorflow
JAX and Tensorflow now work with the Newton Cotes rules and MonteCarlo
Since JAX has a special syntax for in-place changes, I replaced the zeros array initialisation with stacking.
To not break the torch gradients I used stack instead of array for the h values.
MonteCarlo now no longer initialises a tensor full of zeros; I haven't tested the runtime and memory impact of this change yet.
For Tensorflow I replaced the way ravel is called. Tensorflow's ravel only works if numpy behaviour is enabled, otherwise it is missing in the latest tensorflow version.