Extend the test functions to support numpy, torch, jax and tensorflow
I also added a run_example_functions function which in comparison to compute_test_errors additionally returns the functions.
Furthermore, the example functions are not generated on import but when calling run_example_functions. The tests runtime difference due to this change is negligible in comparison to the time required to import a numerical backend.