VEGAS: Exclude estimated variances from the gradient calculation
These variances per iteration are used for the weights calculation in the weighted sum in the final result calculation (the _get_result method), so I think they should be excluded from backpropagation.
I tested the differences by comparing graphs generated by torchviz when the gradient is over parameters of the integrand.
I also added parenthesis in the ih calculation to remove a few nodes in the graph.