RNG: Add an option to save and restore the state with PyTorch
With this the user can change the RNG seed in a VEGAS integrand without changing the sample points.
Measurements showed that the uniform function is ca. 4 µs slower with CPU and 30 µs slower with CUDA when the state is saved and restored, and changing the state may interfere with asynchronous execution, so I disabled the option by default.