Fix JIT compilation with TensorFlow >= 2.14.0
The previous check if code is currently being compiled no longer works with new TensorFlow versions because the `Tensor` type is now called `SymbolicTensor`.

This change adds a helper function to check if code is being compiled for JAX, TensorFlow or PyTorch.
If tf.is_symbolic_tensor() is available, i.e. if the TensorFlow version is high enough,
we use this function to check if code is being compiled.

To avoid inconsistencies between backends,
the check for integration domain values is disabled if code is being compiled with PyTorch even if the check works with PyTorch.